Move logging helpers to homea new logging helper
